Added functionality to drop high or low dice from a roll.  To use, follow this convention:
[n]d[x][<(l|h)m>]

where n is the number of dice, x is the number of sides, l|h is either drop low or high dice, and m is the number to drop.

So for example, typical of D&D character generation:

	rolldice 4d6l1

will roll 4d6 and drop the lowest.

Or using the D&D 5e Advantage mechanic:

	rolldice 2d20h1

will roll 2d20 and drop the highest one, as rolling with disadvantage.